initial implementation

+++ Core/Inc/usart.h
@@ -41,7 +41,7 @@ extern UART_HandleTypeDef huart4;
 void MX_UART4_Init(void);
 
 /* USER CODE BEGIN Prototypes */
-bool command_queue_pop(uint8_t *s);
+bool command_queue_pop(uint32_t *s);
 /* USER CODE END Prototypes */
 
 #ifdef __cplusplus
